final version
This project shows a simple implementation of a PHP Backend. It
contains:

- Speaking urls
- SQL examples
- Authentification & registration example

The index.php is the bootstrap-file which handles the URLs to adress the
right class and method. For example, if we want to call the class
'example' and the dedicated method 'get_random_quote', the corresponding
URL would look like:

http: //www.exampleurl.com/example/get_random_quote